move generic sql package to module package namespace
