+++ src/hakunapi-source-oracle/src/main/java/fi/nls/hakunapi/simple/sdo/sql/ResultSetValueProvider.java
@@ -1,4 +1,4 @@
-package fi.nls.hakunapi.sql;
+package fi.nls.hakunapi.simple.sdo.sql;
 
 import java.nio.charset.StandardCharsets;
 import java.sql.Array;
