+++ src/hakunapi-source-oracle/src/main/java/fi/nls/hakunapi/simple/sdo/sql/SQLSimpleSource.java
@@ -1,4 +1,4 @@
-package fi.nls.hakunapi.sql;
+package fi.nls.hakunapi.simple.sdo.sql;
 
 import java.io.File;
 import java.io.FileInputStream;
