+++ src/hakunapi-source-oracle/src/main/java/fi/nls/hakunapi/simple/sdo/sql/SimpleJoinGraph.java
@@ -1,4 +1,4 @@
-package fi.nls.hakunapi.sql;
+package fi.nls.hakunapi.simple.sdo.sql;
 
 import java.util.ArrayList;
 import java.util.Collections;
