+++ src/hakunapi-source-oracle/src/main/java/fi/nls/hakunapi/simple/sdo/sql/filter/SQLIsNotNull.java
@@ -1,4 +1,4 @@
-package fi.nls.hakunapi.sql.filter;
+package fi.nls.hakunapi.simple.sdo.sql.filter;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -6,14 +6,14 @@ import java.sql.SQLException;
 
 import fi.nls.hakunapi.core.filter.Filter;
 import fi.nls.hakunapi.core.property.HakunaProperty;
-import fi.nls.hakunapi.sql.SQLUtil;
+import fi.nls.hakunapi.simple.sdo.sql.SQLUtil;
 
-public class SQLIsNull implements SQLFilter {
+public class SQLIsNotNull implements SQLFilter {
 
     @Override
     public String toSQL(Filter filter) {
         HakunaProperty prop = filter.getProp();
-        return SQLUtil.toSQL(prop) + " IS NULL";
+        return SQLUtil.toSQL(prop) + " IS NOT NULL";
     }
 
     @Override
