+++ src/hakunapi-source-oracle/src/main/java/fi/nls/hakunapi/simple/sdo/sql/filter/SQLLessThan.java
@@ -1,4 +1,4 @@
-package fi.nls.hakunapi.sql.filter;
+package fi.nls.hakunapi.simple.sdo.sql.filter;
 
 public class SQLLessThan extends SQLComparison {
