+++ src/hakunapi-source-oracle/src/main/java/fi/nls/hakunapi/simple/sdo/sql/filter/SQLNotEqualTo.java
@@ -1,4 +1,4 @@
-package fi.nls.hakunapi.sql.filter;
+package fi.nls.hakunapi.simple.sdo.sql.filter;
 
 public class SQLNotEqualTo extends SQLComparison {
